Added code from interactive_selection_service.py to items_selection_service.py
